+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 
 * Enhancements
   * Support reusable params by `use` DSL within `params`.
+  * readd Maru.Middleware